batch_app/requirements.txt
pyspark==3.5.0
dependency-injector
python-dotenv
requests